LayoutTests/fast/js/dfg-arrayify-when-late-prevent-extensions-expected.txt

Tests that Arraify does good things when Object.preventExtensions() has been called.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is 42
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS foo(o) is void 0
PASS successfullyParsed is true

TEST COMPLETE
